+++ .gitignore
@@ -36,3 +36,4 @@ libs
 data
 build
 python_interface/.idea/
+/mvdet_data/
