+++ C/examples/edgeinference/CMakeLists.txt
