+++ C/examples/edgeinference/edgeinference.cpp
