+++ C/examples/edgeinference/edgeinference_2.json
