+++ C/examples/edgeinference/edgeinference_4.json
