+++ C/examples/edgeinference/edgeinference_7.json
