+++ C/examples/edgeinference/process-frame.cpp
