+++ C/examples/edgeinference/process-frame.hpp
