+++ C/examples/masterworker/masterworker_7.json
