+++ C/examples/mvdet/helpers.cpp
