+++ C/examples/mvdet/helpers.hpp
