+++ C/examples/mvdet/mvdet.cpp
@@ -4,7 +4,7 @@
 #include <filesystem>
 #include <cstdint>
 
-#include "utils/net.hpp"
+#include "C/utils/net.hpp"
 #include "helpers.hpp"
 #include "data_stucts.hpp"
