+++ python_interface/logging.conf
@@ -42,4 +42,4 @@ formatter=customFormatter
 args=(sys.stdout,)
 
 [formatter_customFormatter]
-class=custom.custom_formatter.CustomFormatter
+class=python_interface.custom.custom_formatter.CustomFormatter
